+++ gprof/cg_arcs.h
@@ -15,7 +15,7 @@ typedef struct arc
   {
     Sym *parent;		/* source vertice of arc */
     Sym *child;			/* dest vertice of arc */
-    int count;			/* # of calls from parent to child */
+    unsigned long count;	/* # of calls from parent to child */
     double time;		/* time inherited along arc */
     double child_time;		/* child-time inherited along arc */
     struct arc *next_parent;	/* next parent of CHILD */
@@ -27,7 +27,7 @@ Arc;
 extern unsigned int num_cycles;	/* number of cycles discovered */
 extern Sym *cycle_header;	/* cycle headers */
 
-extern void arc_add PARAMS ((Sym * parent, Sym * child, int count));
+extern void arc_add PARAMS ((Sym * parent, Sym * child, unsigned long count));
 extern Arc *arc_lookup PARAMS ((Sym * parent, Sym * child));
 extern Sym **cg_assemble PARAMS ((void));
 extern Arc **arcs;
